+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpFinallyBlock.cs
@@ -1,0 +1,8 @@
+﻿namespace Intent.Modules.Common.CSharp.Builder;
+
+public class CSharpFinallyBlock : CSharpStatementBlock
+{
+    public CSharpFinallyBlock() : base($"finally")
+    {
+    }
+}